+++ internal/query/client.go
@@ -348,7 +348,8 @@ func (c *Client) QueryRow(ctx context.Context, q string, opts ...options.Execute
 func clientExec(ctx context.Context, pool sessionPool, q string, opts ...options.Execute) (finalErr error) {
 	settings := options.ExecuteSettings(opts...)
 	err := do(ctx, pool, func(ctx context.Context, s *Session) (err error) {
-		streamResult, err := s.execute(ctx, q, settings, withStreamResultTrace(s.trace))
+		streamResult, err := s.execute(ctx, q, settings,
+			withStreamResultTrace(s.trace), withIssuesHandler(settings.IssuesOpts()))
 		if err != nil {
 			return xerrors.WithStackTrace(err)
 		}
@@ -397,7 +398,9 @@ func clientQuery(ctx context.Context, pool sessionPool, q string, opts ...option
 ) {
 	settings := options.ExecuteSettings(opts...)
 	err = do(ctx, pool, func(ctx context.Context, s *Session) (err error) {
-		streamResult, err := s.execute(ctx, q, options.ExecuteSettings(opts...), withStreamResultTrace(s.trace))
+		settingsLocal := options.ExecuteSettings(opts...)
+		streamResult, err := s.execute(ctx, q, settings,
+			withStreamResultTrace(s.trace), withIssuesHandler(settingsLocal.IssuesOpts()))
 		if err != nil {
 			return xerrors.WithStackTrace(err)
 		}