+++ tests/integration/query_execute_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
 	"github.com/ydb-platform/ydb-go-sdk/v3"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/decimal"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/value"
@@ -772,6 +773,7 @@ func TestIssue1785FillDecimalFields(t *testing.T) {
 	})
 }
 
+// https://github.com/ydb-platform/ydb-go-sdk/issues/1872
 func TestIssue1872QueryWarning(t *testing.T) {
 	ctx, cancel := context.WithCancel(xtest.Context(t))
 	defer cancel()
@@ -795,11 +797,17 @@ func TestIssue1872QueryWarning(t *testing.T) {
 	)
 	err = db.Query().Exec(ctx,
 		`create table TestIssue1872QueryWarning 
-		(Id uint64, Amount decimal(22,9), primary key(Id));`,
+		(Id uint64, Amount decimal(22,9) , primary key(Id));`,
+		query.WithParameters(
+			ydb.ParamsBuilder().
+				Param("$p1").Text("test1").
+				Build(),
+		),
 	)
 	require.NoError(t, err)
 
 	t.Run("Query", func(t *testing.T) {
+		var issueList []*Ydb_Issue.IssueMessage
 		q := db.Query()
 		result, err := q.Query(ctx, `
         insert into TestIssue1872QueryWarning (Id, Amount) values (-3, Decimal("3.01",22,9));
@@ -812,16 +820,34 @@ func TestIssue1872QueryWarning(t *testing.T) {
 			),
 			query.WithSyntax(query.SyntaxYQL),
 			query.WithIdempotent(),
+			query.WithIssuesHandler(func(issues []*Ydb_Issue.IssueMessage) {
+				issueList = issues
+				fmt.Printf("len=%d", len(issues))
+			}),
 		)
 		require.NoError(t, err)
-		issues := result.GetIssues()
-		require.Equal(t, 1, len(issues))
-		require.Equal(t, "Type annotation", issues[0].Message)
-		require.Equal(t, 2, len(issues[0].Issues))
-		require.Equal(t, "At function: KiWriteTable!", issues[0].Issues[0].Message)
+		require.Equal(t, 1, len(issueList))
+		require.Equal(t, "Type annotation", issueList[0].Message)
+		require.Equal(t, 2, len(issueList[0].Issues))
+		require.Equal(t, "At function: KiWriteTable!", issueList[0].Issues[0].Message)
 		require.Equal(t,
 			"Failed to convert type: Struct<'Amount':Decimal(22,9),'Id':Int32> to Struct<'Amount':Decimal(22,9)?,'Id':Uint64?>",
-			issues[0].Issues[0].Issues[0].Message)
+			issueList[0].Issues[0].Issues[0].Message)
 		fmt.Printf("%#v", result)
+	})
+
+	t.Run("Exec", func(t *testing.T) {
+		var issueList []*Ydb_Issue.IssueMessage
+		q := db.Query()
+		err := q.Exec(ctx, `
+        insert into TestIssue1872QueryWarning (Id, Amount) values (-7, Decimal("37.01",22,9));
+        `,
+			query.WithIssuesHandler(func(issues []*Ydb_Issue.IssueMessage) {
+				issueList = issues
+				fmt.Printf("len=%d", len(issues))
+			}),
+		)
+		require.NoError(t, err)
+		require.Equal(t, 1, len(issueList))
 	})
 }